src/requirements.txt
Flask>=2.1.0
pg8000==1.29.8
Flask-SQLAlchemy==3.1.1
cloud-sql-python-connector==1.2.4
gunicorn==20.1.0
functions-framework==3.3.0
Werkzeug==2.3.7
# Flask-Migrate==4.0.5
marshmallow-sqlalchemy==0.29.0
flask-marshmallow==0.15.0
firebase-admin==6.2.0
structlog==23.1.0
urllib3<2.0.0 #https://stackoverflow.com/questions/76175361/firebase-authentication-httpresponse-object-has-no-attribute-strict-status
Flask-Migrate==4.0.5
Flask-SocketIO==5.1.1